requirements.txt
beautifulsoup4==4.6.0
cycler==0.10.0
functools32==3.2.3.post2
matplotlib==2.0.2
nltk==3.2.5
numpy==1.13.1
pyenchant==1.6.6
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
subprocess32==3.2.7
